Enable many more busybox applets
Almost every applet is enabled now (359)

Surprisingly the busybox binary is still quite small
and compiles in i486, x86_64, and arm.

I want busybox to be treated differently from all the other programs.

The same binary produced here should go the main SFS, because it's
lightweight and practically complete (it does work in any puppy and boots
to desktop).

Aside from the mandatory symlinks, the build scripts should create
extra symlinks to busybox in /usr/local/bin/busybox/
(that directory should be added to the $PATH as the last dir)

That way users will be surprised to find out that every command they
type it's already there, if it's missing features,
they can still install the full versions afterwards
without causing any problems because the extra busybox symlinks are the last resort.

This could also be useful to identify what other busybox apps work exactly the same
as the full versions, making the busybox app the default one, saving space...