init:  "find_onepupdrv" and "load_onepupdrv" functions
Creates new "find_onepupdrv" and "load_onepupdrv" functions that are used for all 4 sfs files.
Since the exact filename is always known, the actual "finding" is implemented as a simple file test.
The "load_onepupdrv" function is capable of mounting a partition, so any of the 4 sfs's can be on any partition.
All sfs files copied to ram now use a single tmpfs partition.
The size of 'init' is significantly reduced.